+++ src/resources/lang/da_DK/permissionmanager.php
@@ -1,0 +1,29 @@
+<?php
+
+return [
+    /*
+    |--------------------------------------------------------------------------
+    | Permission Manager Language Lines
+    |--------------------------------------------------------------------------
+    |
+    | The following language lines are used for Laravel Backpack - Permission Manager
+    | Author: Frederik Rabøl Madsen <frederik-rm@hotmail.com>
+    |
+    */
+    'name'                  => 'navn',
+    'role'                  => 'rolle',
+    'roles'                 => 'roller',
+    'roles_have_permission' => 'roller der har denne rettighed',
+    'permission_singular'   => 'rettighed',
+    'permission_plural'     => 'rettigheder',
+    'user_singular'         => 'bruger',
+    'user_plural'           => 'brugere',
+    'email'                 => 'E-mail',
+    'extra_permissions'     => 'yderligere rettigheder',
+    'password'              => 'password',
+    'password_confirmation' => 'gentag password',
+    'user_role_permission'  => 'bruger rolle rettigheder',
+    'user'                  => 'bruger',
+    'users'                 => 'brugere',
+
+];